==> code/analysis/infection/main/02_01_cox_analyses_to_run.R <==
##==============================================================================
## Reads in the active analyses table which specifies which 
## analysis to run for the outcome
## Authors: Alexia Sampri, Wen Shi
## Creates a table of all the analyses to run for the outcome
## =============================================================================

## Read in active analyses table and filter to relevant outcome

# expo.rds covariates names not same as data
# change outcome_variable values

active_analyses <- read_rds(paste0(scripts_dir,"active_analyses_",exposure,"_expo.rds"))
active_analyses <- active_analyses %>%dplyr::filter(outcome_variable %in% event_name)

## Select covariates of interest 
covar_names<-str_split(active_analyses$covariates, ";")[[1]]
covar_names<-append(covar_names,"patient_id")
covar_names<-covar_names[!covar_names %in% c("cov_n_age","cov_cat_ethnicity","cov_cat_region","cov_cat_sex")]


##Set which models are required

if(active_analyses$model=="all"){
  mdl=c("mdl_age_sex","mdl_age_agesq_sex","mdl_age_sex_region","mdl_max_adj","mdl_max_adj_agesq","mdl_max_adj_region")
}else if(active_analyses$model=="two") {
  mdl=c("mdl_age_sex","mdl_max_adj_region")
} else {
  mdl=active_analyses$model
}


## Transpose active_analyses to single column so can filter to analysis models to run

analyses_to_run <- as.data.frame(t(active_analyses))
analyses_to_run$subgroup <- row.names(analyses_to_run)
colnames(analyses_to_run) <- c("run","subgroup")
analyses_to_run<- analyses_to_run %>% filter(run=="TRUE" & subgroup != "active" ) 
rownames(analyses_to_run) <- NULL
analyses_to_run <- analyses_to_run %>% select(!run)
analyses_to_run$event=event_name

## Add in  all possible combinations of the subgroups, models and exposure
analyses_to_run <- crossing(analyses_to_run,exposure)

## Add in which covariates to stratify by
analyses_to_run$stratify_by_subgroup=NA
for(i in c("ethnicity","sex")){
  analyses_to_run$stratify_by_subgroup <- ifelse(startsWith(analyses_to_run$subgroup,i),i,analyses_to_run$stratify_by_subgroup)
}

analyses_to_run$stratify_by_subgroup <- ifelse(analyses_to_run$subgroup %in% c("prior_history_expo_TRUE","prior_history_expo_FALSE","adjust_for_expo"),active_analyses$prior_hist_expo_var,analyses_to_run$stratify_by_subgroup)
analyses_to_run$stratify_by_subgroup <- ifelse(is.na(analyses_to_run$stratify_by_subgroup),analyses_to_run$subgroup,analyses_to_run$stratify_by_subgroup)


## Add in relevant subgroup levels to specify which stratum to run for
analyses_to_run$strata <- NA
analyses_to_run$strata <- ifelse(analyses_to_run$subgroup=="main","main",analyses_to_run$strata)

for(i in c("covid_pheno_","agegp_","sex_","ethnicity_","prior_history_expo_")){
  analyses_to_run$strata <- ifelse(startsWith(analyses_to_run$subgroup,i),gsub(i,"",analyses_to_run$subgroup),analyses_to_run$strata)
  
}

analyses_to_run$strata <- ifelse(analyses_to_run$subgroup == "adjust_for_expo",analyses_to_run$stratify_by_subgroup,analyses_to_run$strata)


# add subgroup category

analyses_to_run <- analyses_to_run %>% 
  dplyr::mutate(subgroup_cat = case_when(
    startsWith(subgroup, "agegp") ~ "age",
    startsWith(subgroup, "covid_pheno") ~ "covid_pheno",
    startsWith(subgroup, "ethnicity") ~ "ethnicity",
    startsWith(subgroup, "main") ~ "main",
    startsWith(subgroup, "prior_history") ~ "prior_history",
    startsWith(subgroup, "sex") ~ "sex",
    TRUE ~ as.character(subgroup)))


#Adjust for previous exposure covariate will be run alongside the main analysis

if("adjust_for_expo" %in% analyses_to_run$subgroup){
  adjust_for_expo <- "TRUE"
  adjust_for_expo_covariate <- (analyses_to_run %>% filter(subgroup == "adjust_for_expo") %>% select(stratify_by_subgroup))[[1]]
  analyses_to_run <- analyses_to_run %>% filter(subgroup != "adjust_for_expo")
}else{
  adjust_for_expo <- "FALSE"
}
